HDC Build workflow improvements (#259)
Ticket: https://linear.app/hivemapper/issue/IMCO-196/cicd-improvements-and-bugfixes

Add clean commands for specific packages. Buildroot is not good at picking up when a package needs to be rebuilt, so we force clean the most commonly modified packages.

Also add version.json to deploy message. SHA from the github workflow environment is not very reliable, it only reports the current head of the branch. version.json is extracted from the firmware itself and is always accurate.